pyproject.toml
[tool.poetry]
name = "pilea"
version = "0.1.0"
description="pilea - another opinionated static site generator."
authors = ["hfjn <hi@hfjn.de>"]
license = "MIT"
readme = "README.md"
include = [
  "README.md",
  "LICENSE",
  "pyproject.toml",
  "pilea/input/**/*"
  ]

exclude = ["tests/**/*"]

repository = "https://github.com/hfjn/pilea"

keywords = ["ssg", "python"]

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Topic :: Software Development",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7"
  ]
# esque entrypoint
[tool.poetry.scripts]
pilea = "pilea.cli:pilea"

# Dependencies
[tool.poetry.dependencies]
python = "^3.6"
virtualenv = "*"
click = ">=7.0"
feedgen = "*"
pendulum = "*"
jinja2 = "*"
markdown = "*"
pyyaml = "*"
pygments = "*"
pymdown-extensions = "*"
livereload = "^2.6.1"
Pillow = "^7.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-mock = "^1.11"
pytest-cov = "^2.8"
flake8 = "^3.7"
black = {version = "19.3b0", allows-prereleases = true}
isort = {version = "^4.3", extras = ["pyproject"]}
bandit = "^1.6.2"

[tool.black]
line-length = 119
target_version = ['py36']
include = '\.pyi?$'

[tool.isort]
line_Length = 119
multi_line_output = 3
include_trailing_comma = true
default_section = "THIRDPARTY"
known_first_party = ["pilea", "tests"]

[build-system]
requires = ["poetry>=1.0.0b5"]
build-backend = "poetry.masonry.api"
